requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
yfinance>=0.2.30
matplotlib>=3.7
neuralforecast>=1.7
pyyaml>=6.0
pytest>=7.0
